+++ vscode/src/devtoolsserver.ts
@@ -855,6 +855,7 @@ export class DeveloperToolsManager extends JDEventSource {
         title?: string
         progress: string
         useShell?: boolean
+        verbose?: boolean
         diagnostics?: boolean
         developerMode?: boolean
         internet?: boolean
@@ -882,11 +883,11 @@ export class DeveloperToolsManager extends JDEventSource {
             if (!v) {
                 showErrorWithHelp(
                     "terminal.nodemissing",
-                    "Unable to locate Node.JS v16+."
+                    "Unable to locate Node.JS."
                 )
                 return undefined
             }
-            if (!(v.major >= 16)) {
+            if (!(v.major >= MIN_NODE_VERSION)) {
                 showErrorMessage(
                     "terminal.nodeversion",
                     `Node.JS version outdated, found ${v.major}.${v.minor} but needed v16+.`
@@ -930,6 +931,7 @@ export class DeveloperToolsManager extends JDEventSource {
                     this.lastCreateCliFailed ||
                     (options.useShell ?? !!devToolsConfig.get("shell"))
                 const nodePath = this.nodePath
+                const verbose = options.verbose ?? devToolsConfig.get("verbose")
                 const diagnostics =
                     options.diagnostics ?? jacdacConfig.get("diagnostics")
                 const developerMode =
@@ -943,6 +945,7 @@ export class DeveloperToolsManager extends JDEventSource {
                 if (diagnostics) args.push("--diagnostics", "--verbose")
                 if (developerMode) args.push("--dev")
                 if (internet) args.push("--internet")
+                if (verbose) args.push("--verbose")
                 console.debug(
                     `create terminal: ${useShell ? "shell:" : ""}${
                         cwd.fsPath